@@ -42,67 +91,89 @@
     "webpack",
     "react-hot"
   ],
-  "homepage": "https://github.com/alfg/radicalfm#readme",
+  "homepage": "https://github.com/chentsulin/electron-react-boilerplate#readme",
   "devDependencies": {
-    "asar": "^0.9.0",
-    "babel-core": "^6.3.26",
-    "babel-eslint": "^5.0.0-beta6",
-    "babel-loader": "^6.2.0",
-    "babel-plugin-add-module-exports": "^0.1.2",
-    "babel-plugin-webpack-loaders": "^0.2.1",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.0.1",
-    "babel-preset-stage-0": "^6.3.13",
-    "chai": "^3.3.0",
-    "chromedriver": "^2.19.0",
-    "co-mocha": "^1.1.2",
-    "cross-env": "^1.0.7",
-    "css-loader": "^0.23.1",
-    "del": "^2.0.1",
-    "electron-packager": "^5.2.0",
-    "electron-prebuilt": "^0.36.2",
-    "electron-rebuild": "^1.0.0",
-    "eslint": "^1.3.1",
-    "eslint-config-airbnb": "^2.1.1",
-    "eslint-plugin-react": "^3.13.1",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^0.9.1",
-    "fbjs-scripts": "^0.5.0",
-    "jsdom": "^7.2.2",
+    "asar": "^0.12.4",
+    "babel-core": "^6.21.0",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.2.10",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-dev-expression": "^0.2.1",
+    "babel-plugin-tcomb": "^0.3.24",
+    "babel-plugin-webpack-loaders": "^0.8.0",
+    "babel-polyfill": "^6.20.0",
+    "babel-preset-env": "^1.1.4",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-react-optimize": "^1.0.1",
+    "babel-preset-stage-0": "^6.16.0",
+    "babel-register": "^6.18.0",
+    "babili-webpack-plugin": "^0.0.7",
+    "boiler-room-custodian": "^0.5.0",
+    "chai": "^3.5.0",
+    "concurrently": "^3.1.0",
+    "cross-env": "^3.1.3",
+    "css-loader": "^0.26.1",
+    "devtron": "^1.4.0",
+    "electron": "^1.4.13",
+    "electron-builder": "^10.15.1",
+    "electron-devtools-installer": "^2.0.1",
+    "enzyme": "^2.7.0",
+    "eslint": "^3.12.2",
+    "eslint-config-airbnb": "^13.0.0",
+    "eslint-formatter-pretty": "^1.1.0",
+    "eslint-import-resolver-webpack": "^0.8.0",
+    "eslint-loader": "^1.6.1",
+    "eslint-plugin-compat": "^0.1.0",
+    "eslint-plugin-flowtype": "^2.29.2",
+    "eslint-plugin-flowtype-errors": "^2.0.3",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^2.0.2",
+    "eslint-plugin-mocha": "^4.8.0",
+    "eslint-plugin-promise": "^3.4.0",
+    "eslint-plugin-react": "^6.8.0",
+    "express": "^4.14.0",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "fbjs-scripts": "^0.7.1",
+    "file-loader": "^0.9.0",
+    "flow-bin": "^0.37.4",
+    "flow-typed": "^2.0.0",
+    "html-webpack-plugin": "^2.24.1",
+    "jsdom": "^9.9.1",
+    "json-loader": "^0.5.4",
     "minimist": "^1.2.0",
-    "mocha": "^2.3.0",
-    "node-libs-browser": ">= 0.4.0 <=0.6.0",
-    "react-addons-test-utils": "^0.14.2",
-    "redux-devtools": "^3.0.1",
-    "redux-devtools-dock-monitor": "^1.0.1",
-    "redux-devtools-log-monitor": "^1.0.1",
-    "redux-logger": "^2.3.1",
-    "selenium-webdriver": "^2.48.2",
-    "sinon": "^1.17.2",
-    "style-loader": "^0.13.0",
-    "webpack": "^1.12.1",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.4.1",
-    "webpack-target-electron-renderer": "^0.3.0"
+    "mocha": "^3.2.0",
+    "react-addons-test-utils": "^15.4.1",
+    "redux-logger": "^2.7.4",
+    "sinon": "^1.17.6",
+    "spectron": "^3.4.1",
+    "style-loader": "^0.13.1",
+    "tcomb": "^3.2.16",
+    "url-loader": "^0.5.7",
+    "webpack": "^1.14.0",
+    "webpack-dev-middleware": "^1.9.0",
+    "webpack-hot-middleware": "^2.14.0",
+    "webpack-merge": "^2.0.0",
+    "webpack-validator": "^2.3.0"
   },
   "dependencies": {
-    "electron-debug": "^0.5.1",
-    "font-awesome": "^4.4.0",
-    "history": "^1.17.0",
-    "magnet-uri": "^5.1.3",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2",
-    "react-redux": "^4.0.5",
-    "react-router": "^1.0.3",
-    "redux": "^3.2.1",
-    "redux-promise": "^0.5.0",
-    "redux-thunk": "^1.0.2",
-    "torrent-stream": "^1.0.2"
+    "electron-debug": "^1.1.0",
+    "font-awesome": "^4.7.0",
+    "localforage": "^1.4.3",
+    "magnet-uri": "^5.1.5",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
+    "react-player": "^0.14.1",
+    "react-redux": "^5.0.1",
+    "react-router": "^3.0.0",
+    "react-router-redux": "^4.0.7",
+    "redux": "^3.6.0",
+    "redux-thunk": "^2.1.0",
+    "source-map-support": "^0.4.8",
+    "torrent-stream": "^1.0.3"
   },
   "devEngines": {
-    "node": "4.x || 5.x",
-    "npm": "2.x || 3.x"
+    "node": ">=6.x",
+    "npm": ">=3.x"
   }
 }
