@@ -11,17 +11,13 @@ export default class PlaylistCard extends Component {
     url: PropTypes.string,
   };
 
-  constructor(props) {
-    super(props);
-  }
-
   render() {
     const { title, subtitle, img, url } = this.props;
 
     return (
       <div className={styles.playlistCard}>
         <Link to="/playlist" query={{ magnetUri: url }}>
-            <img className={styles.playlistThumbnail__img} src={img} />
+          <img className={styles.playlistThumbnail__img} src={img} alt={title} />
         </Link>
         <div className={styles.playlistCardContent}>
           <div className={styles.playlistCardContent__title}>{title}</div>
