@@ -1,15 +1,15 @@
-import fs from 'fs';
 import React, { Component, PropTypes } from 'react';
-import { Link } from 'react-router';
 import styles from './Track.module.css';
 
 
 export default class Track extends Component {
   static propTypes = {
-    track: PropTypes.object,
+    track: PropTypes.shape({}),
     setTrackUrl: PropTypes.func.isRequired,
     isPlaying: PropTypes.bool,
-    getFileProgress: PropTypes.func
+    getFileProgress: PropTypes.func,
+    downloadTrack: PropTypes.func,
+    trackId: PropTypes.number
   };
 
   constructor(props) {
@@ -17,23 +17,22 @@ export default class Track extends Component {
     this.handleClick = this.handleClick.bind(this);
     this.state = {
       progress: 0
-    }
+    };
   }
 
   componentDidMount() {
-    var timer = setInterval(() => {
+    const timer = setInterval(() => {
       if (this.state.progress === 100) clearInterval(timer);
       this.props.getFileProgress(this.props.track, (progress) => {
-        this.setState({ progress: progress });
+        this.setState({ progress });
       });
     }, 1000);
-
   }
 
   handleClick() {
     if (this.state.progress !== 100 && !this.props.isPlaying) {
       this.props.downloadTrack(this.props.track);
-      var retry = setInterval(() => {
+      const retry = setInterval(() => {
         if (this.state.progress > 10) {
           this.props.setTrackUrl(this.props.trackId, !this.props.isPlaying);
           clearInterval(retry);
@@ -49,12 +48,12 @@ export default class Track extends Component {
     const { progress } = this.state;
     return (
       <div className={styles.track}>
-          <div className={styles.play} onClick={this.handleClick}>
-            <i className={isPlaying ? "fa fa-pause-circle" : "fa fa-play-circle"}></i>
-          </div>
-          <div className={isPlaying ? `${styles.title} ${styles.playing}` : styles.title}>{track.name}</div>
-          <div className={progress === 100 ? `${styles.trackProgress} ${styles.green}` : styles.trackProgress}>{this.state.progress}%</div>
-          <div className={styles.trackLength}>0:00</div>
+        <div className={styles.play} onClick={this.handleClick}>
+          <i className={isPlaying ? 'fa fa-pause-circle' : 'fa fa-play-circle'} />
+        </div>
+        <div className={isPlaying ? `${styles.title} ${styles.playing}` : styles.title}>{track.name}</div>
+        <div className={progress === 100 ? `${styles.trackProgress} ${styles.green}` : styles.trackProgress}>{this.state.progress}%</div>
+        <div className={styles.trackLength}>0:00</div>
       </div>
     );
   }
