@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
+import { ipcRenderer } from 'electron';
 import { Link } from 'react-router';
 import styles from './SideNav.module.css';
-import { ipcRenderer } from 'electron';
 
 export default class SideNav extends Component {
 
@@ -15,7 +15,6 @@ export default class SideNav extends Component {
     this.state = {
       status: 'offline'
     };
-
   }
 
   componentDidMount() {
@@ -25,7 +24,7 @@ export default class SideNav extends Component {
   }
 
   updateOnlineStatus() {
-    var status = navigator.onLine ? 'online' : 'offline';
+    const status = navigator.onLine ? 'online' : 'offline';
     ipcRenderer.send('online-status-changed', status);
     this.setState({ status });
   }
@@ -33,17 +32,17 @@ export default class SideNav extends Component {
   render() {
     return (
       <div className={styles.sideNav}>
-          <Link to="/" className={styles.logo}>RadicalFM</Link>
+        <Link to="/" className={styles.logo}>RadicalFM</Link>
 
-          <ul className={styles.navLinks}>
-              <li className={styles.title}>Main</li>
-              <Link to="/browse" activeClassName={styles.active}><li><i className="fa fa-search"></i> Browse</li></Link>
-              <Link to="/add" activeClassName={styles.active}><li><i className="fa fa-magnet"></i> Add Playlist</li></Link>
-          </ul>
+        <ul className={styles.navLinks}>
+          <li className={styles.title}>Main</li>
+          <Link to="/browse" activeClassName={styles.active}><li><i className="fa fa-search" /> Browse</li></Link>
+          <Link to="/add" activeClassName={styles.active}><li><i className="fa fa-magnet" /> Add Playlist</li></Link>
+        </ul>
 
-          <ul className={styles.status}>
-              <li><i className="fa fa-signal"></i> <span className={this.state.status}>{this.state.status}</span></li>
-          </ul>
+        <ul className={styles.status}>
+          <li><i className="fa fa-signal" /> <span className={this.state.status}>{this.state.status}</span></li>
+        </ul>
       </div>
     );
   }
