@@ -1,23 +1,16 @@
-import React, { Component, PropTypes } from 'react';
+// @flow
+import React, { Component } from 'react';
 import Player from './Player';
 
 export default class App extends Component {
-  static propTypes = {
-    children: PropTypes.element.isRequired
+  props: {
+    children: HTMLElement
   };
 
   render() {
     return (
       <div>
         {this.props.children}
-        {
-          (() => {
-            if (process.env.NODE_ENV !== 'production') {
-              const DevTools = require('./DevTools');
-              return <DevTools />;
-            }
-          })()
-        }
         <Player />
       </div>
     );
