@@ -6,24 +6,24 @@ localforage.config({
   size: 4980736  // Default.
 });
 
-export default Storage = {
+const storage = {
 
   getItem(key, cb) {
-    localforage.getItem(key).then( (err, val) => {
-      cb(null, val);
-    })
-    .catch((err) => {
+    localforage.getItem(key).then((err, val) =>
+      cb(null, val)
+    ).catch((err) => {
       cb(err);
     });
   },
 
   setItem(key, value, cb) {
-    localforage.setItem(key, value).then( () => {
-      cb(null);
-    })
-    .catch( (err) => {
-      cb(err);
-    });
+    localforage.setItem(key, value).then(() =>
+      cb(null)
+    )
+    .catch((err) =>
+      cb(err)
+    );
   }
+};
 
-}
+export default storage;
