@@ -5,17 +5,17 @@ import path from 'path';
 import _ from 'lodash';
 
 const options = {
-    connections: 100,
-    upload: 10,
-    tmp: '/tmp',
-    verify: true,
-    dht: true,
-    tracker: true,
-    trackers: [
-      'udp://tracker.openbittorrent.com:80',
-    	'udp://open.demonii.com:1337',
-      'udp://tracker.coppersurfer.tk:6969'
-    ]
+  connections: 100,
+  upload: 10,
+  tmp: '/tmp',
+  verify: true,
+  dht: true,
+  tracker: true,
+  trackers: [
+    'udp://tracker.openbittorrent.com:80',
+    'udp://open.demonii.com:1337',
+    'udp://tracker.coppersurfer.tk:6969'
+  ]
 };
 
 const allowedFiles = ['.mp3', '.flac'];  // Must contain .
@@ -48,24 +48,28 @@ export default class TorrentClient extends EventEmitter {
 
   onDownload(piece) {
     console.log('download', piece);
+    this.emit('onDownload');
   }
 
   onUpload(piece) {
     console.log('upload', piece);
+    this.emit('onUpload');
   }
 
   onTorrent(fn) {
     console.log('torrent', fn);
+    this.emit('onTorrent');
   }
 
   onIdle(fn) {
     console.log('idle', fn);
+    this.emit('onIdle');
   }
 
   onWire = (wire) => {
     // console.log('wire', wire);
     this.emit('wire', wire);
-};
+  };
 
   getSwarm = () => {
     return this.engine.swarm;
@@ -74,13 +78,13 @@ export default class TorrentClient extends EventEmitter {
   downloadPlaylist(cb) {
     console.log('downloading playlist');
     this.engine.files.forEach( (file) => {
-        var stream = file.createReadStream();
+      const stream = file.createReadStream();
     });
   }
 
   downloadTrack(file, cb) {
     console.log('downloading track');
-    var stream = file.createReadStream();
+    const stream = file.createReadStream();
   }
 
   getMetadata() {
@@ -90,36 +94,36 @@ export default class TorrentClient extends EventEmitter {
 
   getTracks() {
     console.log('getting tracks');
-    var filtered = _.filter(this.engine.files, (o) => {
+    const filtered = _.filter(this.engine.files, (o) => {
       return _.includes(allowedFiles, path.extname(o.name));
     });
     this.emit('tracks', filtered);
     return filtered;
-  };
+  }
 
   getFileProgress = (file, cb) => {
-    const pieceLength = this.engine.pieceLength;
+    // const pieceLength = this.engine.pieceLength;
 
-    var fileStart = file.offset;
-    var fileEnd = file.offset + file.length;
-    var firstPiece = Math.floor(fileStart / pieceLength);
-    var lastPiece = Math.floor((fileEnd - 1) / pieceLength);
+    const fileStart = file.offset;
+    const fileEnd = file.offset + file.length;
+    // const firstPiece = Math.floor(fileStart / pieceLength);
+    // const lastPiece = Math.floor((fileEnd - 1) / pieceLength);
 
     const url = `${this.engine.path}/${file.path}`;
     fs.stat(url, (err, stat) => {
-      const percent = Math.round((stat.size / (fileEnd - fileStart) * 100));
+      const percent = Math.round(((stat.size / (fileEnd - fileStart)) * 100));
       cb(percent);
     });
   };
 
   getPlaylists(cb) {
     console.log('getting playlists');
-    this.engine.files.forEach( (file) => {
-        var stream = file.createReadStream();
-        stream.on('readable', () => {
-          var output = stream.read();
-          if (output !== null) cb(JSON.parse(output.toString()));
-        });
+    this.engine.files.forEach((file) => {
+      const stream = file.createReadStream();
+      stream.on('readable', () => {
+        const output = stream.read();
+        if (output !== null) cb(JSON.parse(output.toString()));
+      });
     });
     // const allowedFiles = ['.json'];  // Must contain .
     // var filtered = _.filter(this.engine.files, function(o) {
@@ -129,5 +133,3 @@ export default class TorrentClient extends EventEmitter {
     // return filtered;
   }
 }
-
-export default TorrentClient;
