@@ -1,25 +1,43 @@
 import { createStore, applyMiddleware, compose } from 'redux';
-import { persistState } from 'redux-devtools';
 import thunk from 'redux-thunk';
+import { hashHistory } from 'react-router';
+import { routerMiddleware, push } from 'react-router-redux';
+import createLogger from 'redux-logger';
 import rootReducer from '../reducers';
-import DevTools from '../containers/DevTools';
 
-const enhancer = compose(
-  applyMiddleware(thunk),
-  DevTools.instrument(),
-  persistState(
-    window.location.href.match(
-      /[?&]debug_session=([^&]+)\b/
-    )
-  )
+import * as playerActions from '../actions/player';
+
+const actionCreators = {
+  ...playerActions,
+  push,
+};
+
+const logger = createLogger({
+  level: 'info',
+  collapsed: true
+});
+
+const router = routerMiddleware(hashHistory);
+
+// If Redux DevTools Extension is installed use it, otherwise use Redux compose
+/* eslint-disable no-underscore-dangle */
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ?
+  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
+    // Options: http://zalmoxisus.github.io/redux-devtools-extension/API/Arguments.html
+    actionCreators,
+  }) :
+  compose;
+/* eslint-enable no-underscore-dangle */
+const enhancer = composeEnhancers(
+  applyMiddleware(thunk, router, logger)
 );
 
-export default function configureStore(initialState) {
+export default function configureStore(initialState: Object | void) {
   const store = createStore(rootReducer, initialState, enhancer);
 
   if (module.hot) {
     module.hot.accept('../reducers', () =>
-      store.replaceReducer(require('../reducers'))
+      store.replaceReducer(require('../reducers')) // eslint-disable-line global-require
     );
   }
 
